+++ bin/check-jenkins-job-status.rb
@@ -84,7 +84,14 @@ class JenkinsJobChecker < Sensu::Plugin::Check::CLI
   end
 
   def job_status(job_name)
-    jenkins_api_client.job.get_current_build_status(job_name)
+    status = jenkins_api_client.job.get_current_build_status(job_name)
+    # If the job is currently running, get the status of the last build instead
+    if status == 'running'
+      last_build = jenkins_api_client.job.get_current_build_number(job_name) - 1
+      build = jenkins_api_client.job.get_build_details(job_name, last_build)
+      status = build['result'].downcase
+    end
+    status
   rescue
     critical "Error looking up Jenkins job: #{job_name}"
   end